Getting rid of extra files
