+++ fatx/interface.py
@@ -24,26 +24,12 @@ class FatxObject():
 	def registerFilesystem(cls, fs):
 		cls._filesystem = fs
 
-	@staticmethod
-	def splitPath(path):
-		segments = path.split('/')
-		while '' in segments:
-			segments.remove('')
-		return segments
-
-	# Note: static typing with selfreference(for parent) is possible but ugly :(
+	# Note: type hinting with selfreference(for parent) is possible but ugly :(
 	def __init__(self, directoryentry: DirectoryEntry, parent): 
 		self._de = directoryentry
 		self._name = self._de.filename
 		self.attributes = self._de.atr
 		self._parent = parent
-
-	def ls(self, deleted=False):
-		"""
-		if self is a directory, this should list all items
-		in this directory
-		"""
-		raise NotImplementedError("Override this in the subclass")
 
 	def details(self):
 		"""
@@ -58,29 +44,15 @@ class FatxObject():
 		"""
 		return self._parent
 
-	def get(self, path):
-		"""
-		returns the FatxObject for a given path(or file)
-		"""
-		raise NotImplementedError("Override this in the subclass")
-
-	def rename(self, name):
+	def rename(self, name: str):
 		"""
 		renames this object and safes the change to disk
 		"""
-		self._filesystem.rename(self._de, name)
-
-	def exportFile(self):
-		"""
-		returns all bytes belonging to this file
-		"""
-		raise NotImplementedError("Override this in the subclass")
-
-	def importFile(self, data, filename):
-		"""
-		imports a given data-bytearray to filename into this folder
-		"""
-		raise NotImplementedError("Override this in the subclass")
+		try:
+			self._filesystem.rename_object(self._de, name)
+			self._name = self._de.filename
+		except ValueError as e:
+			print(e)
 
 	def delete(self):
 		"""
@@ -89,93 +61,74 @@ class FatxObject():
 		raise NotImplementedError("Override this in the subclass")
 
 	def __str__(self):
-		return self._de.filename
+		return self._name
 
 	def __repr__(self):
 		return str(self.__class__)+ ': ' + str(self)
 
 
 class FileObject(FatxObject):
-
-	def ls(self, deleted=False):
-		raise TypeError("This is a file, not a directory")
-
-	def get(self, path):
-		# since every FatxObject has the scope "/" as root(itself), asking for "/" should yield yourself
-		segments = self.splitPath(path)
-		if len(segments) == 0:
-			return self
-		raise TypeError("This is a file, not a directory(Note: use .exportFile() to extract my contents!")
-
-	def exportFile(self):
-		return self._filesystem.readFile(self._de)
+	def export(self):
+		"""
+		returns all bytes belonging to this file
+		"""
+		return self._filesystem.read_file(self._de)
 
 
 class DirectoryObject(FatxObject):
 	"""
 	 This class represents directorys of the filesystem. 
 	"""
-	def __init__(self, directoryentry: DirectoryEntry, parent: FatxObject): # directorylist: DirectoryEntryList):
+	def __init__(self, directoryentry: DirectoryEntry, parent): # directorylist: DirectoryEntryList):
 		super().__init__(directoryentry, parent)
 
 		# get a list of all files in the subdir, note: You'll get a DirectoryEntryList(DEL) in return
 		# This DEL enables you to append files and writing them to disk
-		self._dl = self._filesystem.openDirectory(directoryentry)
+		self._dl = self._filesystem.open_directory(directoryentry)
 		
 		# Prepare a list of FatxObjects for easy access later on
-		self._elements = []
-		self.createFatxObjectList()
+		self._elements = None
 
-	def createFatxObjectList(self):
-		for i in self._dl.list():
-			if i.atr.DIRECTORY:
-				self._elements.append(DirectoryObject(i, self))
+	def ls(self, deleted=False):
+		"""
+		list all items in this directory
+		"""
+		if self._elements is None:
+			self._elements = self._create_obj_list()
+		return [i for i in self._elements if (not i.attributes.DELETED or deleted)]
+
+	def get(self, name: str):
+		"""
+		returns the FatxObject for a given filename
+		"""
+		if self._elements is None:
+			self._elements = self._create_obj_list()
+
+		for i in self._elements:
+			if i._name == name:
+				return i
+		raise IndexError()
+
+	def import_file(self, filename: str, data: bytes):
+		"""
+		imports a given bytearray to filename into this folder
+		"""
+		try:
+			self._filesystem.import_file(self._dl, filename, data)
+		except ValueError as e:
+			print(e)
+
+	def _create_obj_list(self):
+			elements = []
+			if self._dl is not None:
+				for i in self._dl.list():
+					if i.atr.DIRECTORY:
+						elements.append(DirectoryObject(i, self))
+					else:
+						elements.append(FileObject(i, self))
 			else:
-				self._elements.append(FileObject(i, self))
-
-	def ls(self, path="/", deleted=False):
-		obj = self.get(path)
-		if isinstance(obj, DirectoryObject):
-			return [ i for i in obj._elements if (not i.attributes.DELETED or deleted)]
-		else:
-			raise ValueError("Not a directory")
-
-	def get(self, path):
-		def filterByName(name):
-			for i in self._elements:
-				if i._name == name:
-					return i
-			raise IndexError()
-		segments = self.splitPath(path)
-		if len(segments) > 0:
-			subnames = [i._name for i in self._elements]
-			if segments[0] in subnames:
-				try:
-					return filterByName(segments[0]).get(path.replace(segments[0], '', 1))
-				except ValueError as e:
-					raise e
-			else:
-				raise ValueError("Path not found")
-		else:
-			return self
-
-	def importFile(self, path):
-		filename = os.path.basename(path)
-		if filename in [i._name for i in self._elements]:
-			raise SystemError("File already exists")
-		file = open(path, 'rb')
-		# create new directory entry for the new file
-		size = os.stat(filename).st_size
-		newDE = DirectoryEntry.new(size, filename)
-		# write file data to disk
-		clusterStartID = self._filesystem.writeFile(file, size)
-		newDE.cluster = clusterStartID
-		self._dl.append(newDE)
-		self._filesystem.writeDirectoryEntryList(self._dl)
-		
-		# update our entrys
-		self._elements = []
-		self.createFatxObjectList()
+				print("Warning this Folder errored while reading: "+self._name)
+			return elements
 
 
 class RootObject(DirectoryObject):
@@ -187,8 +140,7 @@ class RootObject(DirectoryObject):
 	def __init__(self, directorylist: DirectoryEntryList):
 		self._parent = self
 		self._dl = directorylist
-		self._elements = []
-		self.createFatxObjectList()
+		self._elements = None
 
 	def details(self):
 		raise TypeError("This is your root!")
