+++ main.py
@@ -1,6 +1,6 @@
 import json, sys
-from fatx import fatx2
-from fatx.interface2 import DirectoryObject
+from fatx import fatx
+from fatx.interface import DirectoryObject
 
 def listfiles(root, path=''):
 	for item in root.ls():
@@ -9,7 +9,7 @@ def listfiles(root, path=''):
 			listfiles(item, path+'/'+item._name)
 
 if __name__ == "__main__":
-	fs = fatx2.Filesystem(sys.argv[1])
+	fs = fatx.Filesystem(sys.argv[1])
 	fs.status()
 	root = fs.root
 	listfiles(root)
